% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic.package-package.r
\docType{package}
\name{tic.packagedocs-package}
\alias{tic.packagedocs}
\alias{tic.packagedocs-package}
\title{tic.packagedocs: Example Package for 'tic'}
\description{
Building a package with packagedocs documentation.
}
\details{
For demonstration of the \pkg{tic} package.
}
\examples{
# Show timestamp
Sys.time()
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/tic.packagedocs}
  \item \url{http://krlmlr.github.io/tic.packagedocs}
  \item Report bugs at \url{https://github.com/krlmlr/tic.packagedocs/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
